✨Personality Slash Command
- ⚒️Also fixes crashing when webhook log is greater than discords max embed size
